[WIP] inversify/InversifyJS/issues/279
